added DateMath

+++ src/App.tsx
@@ -8,6 +8,7 @@ import { YearSelector } from "./components/YearSelector";
 // import style from "./App.module.css";
 import "./themes/ark-ui/ark-ui.css";
 
+
 const App: Component = () => {
   const [value, setValue] = createSignal("");
   const [select, setSelect] = createSignal("");
@@ -34,7 +35,7 @@ const App: Component = () => {
       ></div>
 
       <DatePickerGroup
-        weekDaysType="single"
+        weekDaysType="double"
         type={"range"}
         value={singleCustomDate}
         setValue={setSingleCustomDate}
